# vercel-api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: PostCrafter API for GPT Actions
  description: |
    A comprehensive API for creating and managing WordPress posts through ChatGPT GPT Actions.
    
    ## Features
    - **Draft/Publish Control**: Choose whether to save posts as drafts or publish immediately
    - **Status Management**: Full post status lifecycle management
    - **WordPress Integration**: Direct integration with WordPress REST API
    - **SEO Support**: Yoast SEO fields integration
    - **Image Handling**: URL and base64 image upload support
    - **Security**: API key authentication and rate limiting
    
    ## Post Status Options
    - **draft**: Save post as draft for review and editing
    - **publish**: Publish post immediately to make it live
    - **private**: Save as private post (visible only to authorized users)
    
    ## Usage in ChatGPT
    When using this GPT Action, you can specify the post status to control publication:
    - "Create a draft post about..." - Saves as draft
    - "Publish a post about..." - Publishes immediately
    - "Create a private post about..." - Saves as private
    
  version: 2.0.0
  contact:
    name: PostCrafter Support
    url: https://github.com/postcrafter/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-vercel-deployment.vercel.app
    description: Production server
  - url: https://your-staging-deployment.vercel.app
    description: Staging server
security:
  - ApiKeyAuth: []
paths:
...
  /api/posts/status:
    put:
      summary: Update Post Status
      description: |
        Update the status of an existing post with validation and audit trail.
        
        ## Status Transitions
        - `draft` → `publish`: Publish a draft post
        - `publish` → `draft`: Revert published post to draft
        - `draft` → `private`: Make draft private
        - `private` → `publish`: Publish private post
        
      operationId: updatePostStatus
      tags:
        - Posts
        - Status Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostStatusUpdateRequest'
      responses:
        '200':
          description: Post status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostStatusUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get your API key from your PostCrafter dashboard.
        
        Example: `X-API-Key: pc_1234567890abcdef`
  schemas:
    PostStatus:
      type: string
      enum: [draft, publish, private]
      description: |
        Post publication status:
        - `draft`: Post is saved but not published (visible in admin only)
        - `publish`: Post is published and publicly visible
        - `private`: Post is published but only visible to authorized users
      example: "draft"
...
    PostStatusUpdateRequest:
      type: object
      description: Request to update post status
      required: [post_id, new_status]
      properties:
        post_id:
          type: integer
          minimum: 1
          description: ID of the post to update
        new_status:
          $ref: '#/components/schemas/PostStatus'
        reason:
          type: string
          maxLength: 500
          description: Reason for status change
          nullable: true
        changed_by:
          type: string
          maxLength: 255
          description: User making the change
          nullable: true
...
    PostStatusUpdateResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            post_id:
              type: integer
            previous_status:
              $ref: '#/components/schemas/PostStatus'
            new_status:
              $ref: '#/components/schemas/PostStatus'
            timestamp:
              type: string
              format: date-time
            wordpress_post_id:
              type: integer
              nullable: true
        error:
          type: string
          nullable: true
...
    ErrorDetails:
      type: object
      description: Error information
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Post title is required"
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Additional error details
          nullable: true
        request_id:
          type: string
          description: Request ID for tracking
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/ErrorDetails'
          examples:
            validation_error:
              summary: Validation Error
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid post data"
                  details: ["Title is required", "Content must be at least 10 characters"]
            invalid_status:
              summary: Invalid Status
              value:
                success: false
                error:
                  code: "STATUS_VALIDATION_ERROR"
                  message: "Invalid post status"
                  details: "Status must be one of: draft, publish, private"
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      headers:
        WWW-Authenticate:
          description: Authentication method required
          schema:
            type: string
            example: "ApiKey"
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/ErrorDetails'
          example:
            success: false
            error:
              code: "AUTHENTICATION_ERROR"
              message: "Invalid or missing API key"
              details: "Please provide a valid API key in the X-API-Key header"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/ErrorDetails'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Post not found"
              details: "No post found with the specified ID"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
            example: 100
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
            example: 1640995800
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/ErrorDetails'
          example:
            success: false
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit exceeded"
              details: "Please wait 60 seconds before making another request"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/ErrorDetails'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              details: "Please try again later or contact support"
...